// back-end/src/vip-vsp/types/gateway-events.ts
enum GatewayEvents {
  IMPORT_VIP = 'import-vip',
  IMPORT_VIP_SUCCESS = 'import-vip-success',
  IMPORT_VIP_FAILED = 'import-vip-failed',

  LOAD_VIP = 'load-vip',
  LOAD_VIP_SUCCESS = 'load-vip-success',
  LOAD_VIP_FAILED = 'load-vip-failed',

  UPDATE_VIP = 'update-vip',
  UPDATE_VIP_SUCCESS = 'update-vip-success',
  UPDATE_VIP_FAILED = 'update-vip-failed',
}
export default GatewayEvents;